requirements.txt
openai
faiss-cpu
numpy
python-dotenv
neo4j
networkx
docling
sentence-transformers
flask
graphrag
lancedb